More separation of interface and implementation. Dependencies of most typechecker parts are (where possible) on interfaces. Moved interfaces to separate files to help with incremental compilation. A change in the DefaultTyper (for example) now causes a smaller ripple in the sea of compiled files.
