+++ src/compiler/scala/reflect/macros/runtime/MacroRuntimesApi.scala
@@ -1,0 +1,10 @@
+package scala.reflect.macros.runtime
+
+import scala.tools.nsc.typechecker
+
+trait MacroRuntimes {
+  self: typechecker.Macros =>
+
+  // used externally
+  type MacroRuntime = MacroArgs => Any
+}
