+++ src/compiler/scala/tools/nsc/typechecker/Macros.scala
@@ -17,65 +17,6 @@ import scala.tools.reflect.FastTrack
 import scala.runtime.ScalaRunTime
 import Fingerprint._
 
-trait Macros extends MacroRuntimes {
-  self: Globals with
-  StdAttachments with
-  // needed for FastTrack
-  Analyzer =>
-    
-  import global._
-
-  def standardMacroExpand(typer: Typer, expandee: Tree, mode: Mode, pt: Type): Tree
-  def standardTypedMacroBody(typer: Typer, macroDdef: DefDef): Tree
-  def standardMacroArgs(typer: Typer, expandee: Tree): MacroArgs
-  
-  case class MacroArgs(c: MacroContext, others: List[Any])
-  
-  def macroExpandWithCallbacks(typer: Typer, expandee:Tree, mode: Mode, pt: Type, listener: MacroExpanderListener): Tree
-  
-  trait MacroExpanderListener {
-    def onSuccess: Option[Tree => Tree] = None
-    def onFallback: Option[Tree => Tree] = None
-    def onSuppressed: Option[Tree => Tree] = None
-    def onDelayed: Option[Tree => Tree] = None
-    def onSkipped: Option[Tree => Tree] = None
-    def onFailure: Option[Tree => Tree] = None
-  }
-  
-  private[scala] def computeMacroDefTypeFromMacroImplRef(macroDdef: DefDef, macroImplRef: Tree): Type
-  private[scala] def openMacros:List[MacroContext]
-  private[scala] def defaultMacroClassloader: ClassLoader
-  private[scala] def enclosingMacroPosition:Position
-  private[scala] def findMacroClassLoader(): ClassLoader
-  private[scala] def untypeMetalevel(tp: Type): Type
-  private[scala] def increaseMetalevel(pre: Type, tp: Type): Type
-  private[scala] def transformTypeTagEvidenceParams(macroImplRef: Tree, transform: (Symbol, Symbol) => Symbol): List[List[Symbol]]
-  private[scala] def loadMacroImplBinding(macroDef: Symbol): Option[MacroImplBinding]
-  private[scala] def fastTrack:FastTrack[self.type]
-  
-  private[typechecker] def macroLogVerbose(msg: => Any):Unit
-  private[typechecker] def globalSettings:Settings
-  private[typechecker] def isBlackbox(macroDef: Symbol): Boolean
-  private[typechecker] def notifyUndetparamsInferred(undetNoMore: List[Symbol], inferreds: List[Type]): Unit
-  private[typechecker] def standardIsBlackbox(macroDef: Symbol): Boolean
-  private[typechecker] def standardMacroRuntime(expandee: Tree): MacroRuntime
-  private[typechecker] def hasPendingMacroExpansions:Boolean
-  private[typechecker] def macroExpand(typer: Typer, expandee: Tree, mode: Mode, pt: Type): Tree
-  private[typechecker] def macroExpandAll(typer: Typer, expandee: Tree): Tree
-  private[typechecker] def notifyUndetparamsAdded(newUndets: List[Symbol]): Unit
-  private[typechecker] def typedMacroBody(typer: Typer, macroDdef: DefDef): Tree
-  
-  private[scala] trait MacroImplBinding {
-    private[typechecker] def isBlackbox:Boolean
-    private[scala] def isBundle:Boolean
-    private[typechecker] def signature: List[List[Fingerprint]]
-    private[typechecker] def targs: List[Tree]
-    private[typechecker] def is_??? : Boolean
-    private[scala] def className: String
-    private[scala] def methName: String
-  }
-}
-
 /**
  *  Code to deal with macros, namely with:
  *    * Compilation of macro definitions
@@ -106,16 +47,16 @@ trait DefaultMacros extends Macros with DefaultMacroRuntimes with Traces with He
   self: Globals 
     with Contexts
     with Namers
-    with DefaultTypers
-    with DefaultInfer
+    with Typers
+    with Infer
     with Implicits
     with EtaExpansion
     with SyntheticMethods
     with Unapplies
-    with DefaultNamesDefaults
-    with DefaultTypeDiagnostics
+    with NamesDefaults
+    with TypeDiagnostics
     with DefaultContextErrors
-    with DefaultStdAttachments
+    with StdAttachments
     with AnalyzerPlugins
     // for FastTrack
     with Analyzer =>
