+++ src/compiler/scala/tools/nsc/typechecker/StdAttachments.scala
@@ -1,50 +1,12 @@
 package scala.tools.nsc
 package typechecker
 
-trait StdAttachments {
-  self: Globals =>
-    
-  import global._
-
-  def markMacroImplRef(tree: Tree): Tree
-  
-  private[typechecker] def macroExpanderAttachment(tree: Tree): MacroExpanderAttachment
-  
-  private[scala] trait MacroExpanderAttachment {
-    private[typechecker] def original: Tree
-    private[typechecker] def desugared: Tree
-  }
-  private[scala] object MacroExpanderAttachment {
-    private[scala] def unapply(m:MacroExpanderAttachment):Option[(Tree, Tree)] = 
-      Option(m).map(m => (m.original, m.desugared))
-  }
-  
-  private[scala] trait MacroExpansionAttachment {
-    private[typechecker] def expandee: Tree 
-    private[typechecker] def expanded: Any
-  }
-  private[scala] object MacroExpansionAttachment {
-    private[scala] def unapply(m:MacroExpansionAttachment):Option[(Tree, Any)] = 
-      Option(m).map(m => (m.expandee, m.expanded))
-  }
-  
-  private[tools] type MacroContext = UnaffiliatedMacroContext { val universe: self.global.type }
-  
-  /** Carries information necessary to expand the host tree.
-   *  At times we need to store this info, because macro expansion can be delayed until its targs are inferred.
-   *  After a macro application has been successfully expanded, this attachment is destroyed.
-   */
-  private[typechecker] type UnaffiliatedMacroContext = scala.reflect.macros.contexts.Context
-}
-
 trait DefaultStdAttachments extends StdAttachments {
   //self: Analyzer =>
   self: Globals with 
   Contexts =>
 
   import global._
-
-  case class MacroRuntimeAttachment(delayed: Boolean, typerContext: Context, macroContext: Option[MacroContext])
 
   /** Scratchpad for the macro expander, which is used to store all intermediate data except the details about the runtime.
    */
@@ -138,13 +100,6 @@ trait DefaultStdAttachments extends StdAttachments {
         case _                => false
       })
     )
-
-  /** After being synthesized by the parser, primary constructors aren't fully baked yet.
-   *  A call to super in such constructors is just a fill-me-in-later dummy resolved later
-   *  by `parentTypes`. This attachment coordinates `parentTypes` and `typedTemplate` and
-   *  allows them to complete the synthesis.
-   */
-  case class SuperArgsAttachment(argss: List[List[Tree]])
 
   /** Convenience method for `SuperArgsAttachment`.
    *  Compared with `MacroRuntimeAttachment` this attachment has different a usage pattern,
