+++ src/compiler/scala/tools/nsc/typechecker/Tags.scala
@@ -13,8 +13,6 @@ private[typechecker] trait Tags {
   trait Tag {
     self: Typer =>
 
-    private val runDefinitions = currentRun.runDefinitions
-
     private def resolveTag(pos: Position, taggedTp: Type, allowMaterialization: Boolean) = enteringTyper {
       def wrapper (tree: => Tree): Tree = if (allowMaterialization) (context.withMacrosEnabled[Tree](tree)) else (context.withMacrosDisabled[Tree](tree))
       wrapper(inferImplicit(
