+++ src/compiler/scala/tools/nsc/typechecker/TypersTracking.scala
@@ -14,6 +14,7 @@ import Mode._
 private[typechecker] trait TypersTracking {
   //self: Analyzer =>
   self: Globals with
+  Typers with
   Contexts =>
 
   import global._
@@ -51,7 +52,7 @@ private[typechecker] trait TypersTracking {
     s"($marker$undet_s: ${context.siteString}$owner_long_s) $implicits_s"
   }
 
-  object typingStack {
+  object typingStack extends TypingStackObject {
     val out = new java.io.PrintWriter(System.err, true)
 
     // TODO - account for colors so the color of a multiline string
