+++ src/compiler/scala/tools/nsc/typechecker/Unapplies.scala
@@ -8,22 +8,6 @@ package typechecker
 
 import symtab.Flags._
 
-trait Unapplies {
-  self: Globals =>
-    
-  import global._
-  
-  def companionModuleDef(cdef: ClassDef, parents: List[Tree] = Nil, body: List[Tree] = Nil): ModuleDef
-  
-  private[typechecker] def directUnapplyMember(tp: Type): Symbol
-  private[typechecker] def factoryMeth(mods: Modifiers, name: TermName, cdef: ClassDef): DefDef
-  private[typechecker] def caseModuleDef(cdef: ClassDef): ModuleDef
-  private[typechecker] def caseModuleApplyMeth(cdef: ClassDef): DefDef
-  private[typechecker] def caseModuleUnapplyMeth(cdef: ClassDef): DefDef
-  private[typechecker] def caseClassCopyMeth(cdef: ClassDef): Option[DefDef]
-  private[typechecker] def unapplyMember(tp: Type): Symbol
-}
-
 /*
  *  @author  Martin Odersky
  *  @version 1.0
